# backend/docs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ArchiTrack API",
    "version": "1.0.0",
    "description": "ArchiTrack Backend REST API Documentation",
    "contact": {
      "name": "ArchiTrack Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.architrack.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication token"
      }
    },
...
  },
...
  "paths": {
...
    "/api/site-surveys/{id}/images": {
      "get": {
        "summary": "画像一覧取得",
        "description": "現場調査に紐付く画像の一覧を表示順序でソートして取得",
        "tags": [
          "Survey Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "現場調査ID"
          }
        ],
        "responses": {
          "200": {
            "description": "画像一覧"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "現場調査が見つからない"
          }
        }
      },
      "post": {
        "summary": "画像アップロード",
        "description": "現場調査に画像をアップロード（複数ファイル対応）",
        "tags": [
          "Survey Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "現場調査ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "アップロード成功"
          },
          "400": {
            "description": "バリデーションエラー"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "現場調査が見つからない"
          },
          "413": {
            "description": "ファイルサイズ超過"
          }
        }
      }
    },
    "/api/site-surveys/{id}/images/order": {
      "put": {
        "summary": "画像順序変更",
        "description": "画像の表示順序を一括変更",
        "tags": [
          "Survey Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "現場調査ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orders"
                ],
                "properties": {
                  "orders": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "id",
                        "order"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "order": {
                          "type": "integer",
                          "minimum": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "順序変更成功"
          },
          "400": {
            "description": "バリデーションエラー"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "現場調査または画像が見つからない"
          }
        }
      }
    },
    "/api/site-surveys/images/{imageId}": {
      "delete": {
        "summary": "画像削除",
        "description": "画像と関連する注釈データを削除",
        "tags": [
          "Survey Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "画像ID"
          }
        ],
        "responses": {
          "204": {
            "description": "削除成功"
          },
          "400": {
            "description": "無効な画像ID"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "画像が見つからない"
          }
        }
      },
      "patch": {
        "summary": "画像メタデータ更新",
        "description": "画像のコメントと報告書出力フラグを更新",
        "tags": [
          "Survey Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "画像ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "comment": {
                    "type": "string",
                    "nullable": true,
                    "maxLength": 2000,
                    "description": "写真コメント（最大2000文字、nullでクリア）"
                  },
                  "includeInReport": {
                    "type": "boolean",
                    "description": "報告書出力フラグ"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "comment": {
                      "type": "string",
                      "nullable": true
                    },
                    "includeInReport": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "画像が見つからない"
          }
        }
      }
    },
    "/api/site-surveys/images/{imageId}/thumbnail": {
      "patch": {
        "summary": "サムネイル更新",
        "description": "注釈付き画像からサムネイルを再生成して更新",
        "tags": [
          "Survey Images"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "imageId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "画像ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "imageData"
                ],
                "properties": {
                  "imageData": {
                    "type": "string",
                    "description": "Base64エンコードされた画像データ（data:image/...形式）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "サムネイル更新成功"
          },
          "400": {
            "description": "無効な画像データ"
          },
          "401": {
            "description": "認証エラー"
          },
          "403": {
            "description": "権限不足"
          },
          "404": {
            "description": "画像が見つからない"
          }
        }
      }
    },
...
  }
}